Show new entries on top
Resolves https://github.com/johshoff/find-tempo/issues/1

The reason this was difficult before was that I was using d3 for rendering the
table. The code is slightly larger now, but at least I'm in control.

Probably could have used Angular or something similar to make fewer updates to
the DOM and also have less code, but not doing that in the spirit of getting
things done.